// src/slices/types.ts
export type AppThunk<ReturnType = void> = ThunkAction<ReturnType, State, unknown, AnyAction>;

export type DeserializedPoolVault = DeserializedPool & DeserializedCakeVault;

export interface BigNumberToJson {
  type: "BigNumber";
  hex: string;
}

export type SerializedBigNumber = string | null;
// ...
export enum VaultKey {
  CakeVault = "cakeVault",
}

interface CorePoolProps {
  startBlock?: number;
  endBlock?: number;
  apr?: number;
  rawApr?: number;
  stakingTokenPrice?: number;
  earningTokenPrice?: number;
  vaultKey?: VaultKey;
}
// ...
export interface SerializedPool extends SerializedPoolConfig, CorePoolProps {
  totalStaked?: SerializedBigNumber;
  stakingLimit?: SerializedBigNumber;
  numberBlocksForUserLimit?: number;
  profileRequirement?: {
    required: boolean;
    thresholdPoints: SerializedBigNumber;
  };
  userData?: {
    allowance: SerializedBigNumber;
    stakingTokenBalance: SerializedBigNumber;
    stakedBalance: SerializedBigNumber;
    pendingReward: SerializedBigNumber;
  };
}

// export interface Profile {
//   userId: number
//   points: number
//   teamId: number
//   collectionAddress: string
//   tokenId: number
//   isActive: boolean
//   username: string
//   nft?: NftToken
//   team?: Team
//   hasRegistered: boolean
// }

// Slices states

export interface SerializedFarmsState {
  data: SerializedFarm[];
  loadArchivedFarmsData: boolean;
  userDataLoaded: boolean;
  loadingKeys: Record<string, boolean>;
  poolLength?: number;
  regularCakePerBlock?: number;
}

export interface DeserializedFarmsState {
  data: DeserializedFarm[];
  loadArchivedFarmsData: boolean;
  userDataLoaded: boolean;
  poolLength?: number;
  regularCakePerBlock?: number;
}

export interface SerializedVaultFees {
  performanceFee: number | null;
  withdrawalFee: number | null;
  withdrawalFeePeriod: number | null;
}

export interface DeserializedVaultFees extends SerializedVaultFees {
  performanceFeeAsDecimal: number;
}

interface SerializedVaultUser {
  isLoading: boolean | null;
  userShares: SerializedBigNumber | null;
  cakeAtLastUserAction: SerializedBigNumber | null;
  lastDepositedTime: string | null;
  lastUserActionTime: string | null;
}

export interface SerializedLockedVaultUser extends SerializedVaultUser {
  lockStartTime: string | null;
  lockEndTime: string | null;
  userBoostedShare: SerializedBigNumber | null;
  locked: boolean | null;
  lockedAmount: SerializedBigNumber | null;
  currentPerformanceFee: SerializedBigNumber | null;
  currentOverdueFee: SerializedBigNumber | null;
}
// ...
export interface SerializedCakeVault {
  totalShares?: SerializedBigNumber | null;
  totalLockedAmount?: SerializedBigNumber | null;
  pricePerFullShare?: SerializedBigNumber | null;
  totalCakeInVault?: SerializedBigNumber | null;
  fees?: SerializedVaultFees | null;
  userData?: SerializedLockedVaultUser | null;
}

export interface PoolsState {
  data: SerializedPool[];
  cakeVault: SerializedCakeVault;
  userDataLoaded: boolean;
}
// ...
export interface PredictionsState {
  status: PredictionStatus;
  isLoading: boolean;
  isHistoryPaneOpen: boolean;
  chartView: PredictionsChartView;
  isChartPaneOpen: boolean;
  isFetchingHistory: boolean;
  historyFilter: HistoryFilter;
  currentEpoch: number;
  intervalSeconds: number;
  minBetAmount: string;
  bufferSeconds: number;
  history: Bet[];
  totalHistory: number;
  currentHistoryPage: number;
  hasHistoryLoaded: boolean;
  rounds?: RoundData;
  ledgers?: LedgerData;
  claimableStatuses: {
    [key: string]: boolean;
  };
  leaderboard: {
    selectedAddress: string;
    loadingState: FetchStatus;
    filters: LeaderboardFilter;
    skip: number;
    hasMoreResults: boolean;
    addressResults: {
      [key: string]: PredictionUser;
    };
    results: PredictionUser[];
  };
}
// ...
export interface LotteryState {
  currentLotteryId: string;
  maxNumberTicketsPerBuyOrClaim: string;
  isTransitioning: boolean;
  currentRound: LotteryResponse & { userTickets?: LotteryRoundUserTickets };
  lotteriesData?: LotteryRoundGraphEntity[];
  userLotteryData?: LotteryUserGraphEntity;
}
// ...
export interface State {
  farms: SerializedFarmsState;
  farmsV1: SerializedFarmsState;
  pools: PoolsState;
  predictions: PredictionsState;
  lottery: LotteryState;
  // nftMarket: NftMarketState
}